+++ src/tests/test.cpp
@@ -193,8 +193,8 @@ void SampleTest::run()
         for (int x = 0; x < pdf.width(); x++)
         {
             float accum = 0.f;
-            for (int sx = 0; sx < super_samples; ++sx)
-                for (int sy = 0; sy < super_samples; ++sy)
+            for (int sx = 0; sx < (int)super_samples; ++sx)
+                for (int sy = 0; sy < (int)super_samples; ++sy)
                 {
                     Vec3f dir =
                         pixel_to_sample(Vec2f{x + (sx + 0.5f) / super_samples, y + (sy + 0.5f) / super_samples});
